Extract connectToDb
