+++ server/utils/db.ts
@@ -1,20 +1,5 @@
-import { UserSchema } from '../models/user.model.ts';
-import { RoomSchema } from '../models/room.model.ts';
-import { MongoClient, ObjectId, config } from '../deps.ts';
-config({ path: './.env' });
-
-const connectToDb = async () => {
-  const DB_URL = Deno.env.get('DB_URL');
-  if (!DB_URL) throw new Error('Problem importing from .env');
-  const client = new MongoClient();
-  await client.connect(DB_URL);
-  const db = client.database('devs_playing_poker');
-
-  const rooms = db.collection<RoomSchema>('rooms');
-  const users = db.collection<UserSchema>('users');
-
-  return { rooms, users };
-};
+import { ObjectId } from '../deps.ts';
+import connectToDb from './connectToDb.ts';
 
 export const { rooms, users } = await connectToDb();
